pyproject.toml
[project]
name = "neuro-classifier"
version = "1.0.0"
description = "Классификация сигналов EPSP/PS с помощью нейросетей"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "tensorflow>=2.13",
    "matplotlib>=3.7",
    "scikit-learn>=1.3",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]